[refactor] 파일 위치 변경 및 GameResultStatus Enum 폴더로 분리
